+++ CHANGELOG.md
@@ -16,7 +16,7 @@ Version `v1.5.1 (15 Nov'23)`
   - CHANGE: don't close "Projects/Files/No name"
   - CHANGE: Help/About/Packages tab
   - CHANGE: alited.ini
-  - CHANGE: packages: e_menu/apave 4.2.0, tkcc, aloupe
+  - CHANGE: packages: e_menu/apave 4.2.0, tkcc, aloupe 1.0, baltip 1.5.5, playtkl 1.0.3
 
 
 Version `v1.4.6 (25 Oct'23)`